Fix precision issue of pow(int32 tensor, float scalar).
Currently we are casting float scalar to int32 tensor (to
match input1). This is incorrect as it would do
power in int and get incorrect results.

Fixed version will cast both to float and do the math in
float. The return value will be float tensor instead of
int tensor